Rework for Transaction Pagination Table
