+++ lib/blocks/widgets/custom_block_widgets.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_annulus/shared/constants/strings.dart';
 import 'package:flutter_annulus/shared/utils/theme_color.dart';
-import 'package:flutter_annulus/transactions/models/transaction.dart';
-import 'package:flutter_annulus/transactions/widgets/custom_transaction_widgets.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:responsive_framework/responsive_breakpoints.dart';
 import 'package:vrouter/vrouter.dart';
@@ -93,129 +91,5 @@ class CustomTabBar extends HookConsumerWidget {
         ),
       ],
     );
-  }
-}
-
-/// This is a custom transaction table that is shown on the block details
-class CustomPaginatedTable extends StatelessWidget {
-  const CustomPaginatedTable({
-    super.key,
-    required this.source,
-    required int rowsPerPage,
-    required this.colorMode,
-  }) : _rowsPerPage = rowsPerPage;
-
-  final TableDataSource source;
-  final int _rowsPerPage;
-  final ThemeMode colorMode;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top: 20, left: 20.0, right: 0.0),
-      child: SizedBox(
-        width: double.infinity,
-        child: SingleChildScrollView(
-          child: Theme(
-              data: Theme.of(context).copyWith(
-                cardColor: getSelectedColor(colorMode, 0xFFFEFEFE, 0xFF282A2C),
-              ),
-              child: PaginatedDataTable(
-                source: source,
-                showFirstLastButtons: true,
-                rowsPerPage: _rowsPerPage,
-                dataRowHeight: 80,
-                columnSpacing: 0,
-                availableRowsPerPage: const [1, 5, 10, 50],
-                onRowsPerPageChanged: (newRowsPerPage) {
-                  if (newRowsPerPage != null) {}
-                },
-                onPageChanged: (int? n) {},
-                columns: const [
-                  DataColumn(
-                    label: Padding(
-                      padding: EdgeInsets.only(right: 60.0),
-                      child: SizedBox(
-                        child: TableHeaderText(name: Strings.tableHeaderTxnHashId),
-                      ),
-                    ),
-                  ),
-                  DataColumn(
-                    label: Padding(
-                      padding: EdgeInsets.only(right: 60.0),
-                      child: TableHeaderText(name: Strings.tableHeaderBlock),
-                    ),
-                  ),
-                  DataColumn(
-                    label: Padding(
-                      padding: EdgeInsets.only(right: 60.0),
-                      child: TableHeaderText(name: Strings.tableHeaderType),
-                    ),
-                  ),
-                  DataColumn(
-                    label: Padding(
-                      padding: EdgeInsets.only(right: 60.0),
-                      child: TableHeaderText(name: Strings.tableHeaderFee),
-                    ),
-                  ),
-                ],
-              )),
-        ),
-      ),
-    );
-  }
-}
-
-/// This is a custom data source for the transaction table
-class TableDataSource extends DataTableSource {
-  TableDataSource(this.data, this.context, this.clr);
-
-  BuildContext context;
-  List<Transaction> data;
-  Color clr;
-
-  @override
-  DataRow? getRow(int index) {
-    final row = data[index];
-    if (index < data.length) {
-      return DataRow(
-          color: MaterialStateProperty.all(clr),
-          onLongPress: () {
-            // Add what you want to do on tap
-          },
-          cells: <DataCell>[
-            const DataCell(TransactionColumnText(
-              textTop: 'row.',
-              textBottom: "49 ${Strings.secAgo}",
-            )),
-            DataCell(TransactionColumnText(
-              textTop: '${Strings.height}: ${row.block.height}',
-              textBottom: '${Strings.slot}: ${row.block.slot}',
-            )),
-            DataCell(TransactionColumnText(
-                textTop: '${row.amount} ${Strings.topl}', textBottom: '${row.quantity} ${Strings.bobs}')),
-            DataCell(TransactionColumnText(
-              textTop: '${row.transactionFee} ${Strings.feeAcronym}',
-              textBottom: "",
-              isBottomTextRequired: false,
-            )),
-          ]);
-    } else {
-      return null;
-    }
-  }
-
-  @override
-  bool get isRowCountApproximate => false;
-
-  @override
-  int get rowCount => data.length;
-
-  @override
-  int get selectedRowCount => 0;
-
-  @override
-  void notifyListeners() {
-    super.notifyListeners();
   }
 }
