+++ lib/transactions/sections/transaction_row_item.dart
@@ -124,89 +124,3 @@ class TransactionTableRow extends HookConsumerWidget {
             ]));
   }
 }
-
-/// Data source class for obtaining row data for PaginatedDataTable.
-class RowDataSource extends DataTableSource {
-  final BuildContext context;
-  final List<Transaction> data;
-  final Color clr;
-  RowDataSource({
-    required this.data,
-    required this.context,
-    required this.clr,
-  });
-
-  @override
-  DataRow? getRow(int index) {
-    print('QQQQ tracations ${data.length}');
-    final isDesktop = ResponsiveBreakpoints.of(context).equals(DESKTOP);
-    final isTablet = ResponsiveBreakpoints.of(context).equals(TABLET);
-
-    final row = data[index];
-    if (index < data.length) {
-      return DataRow(
-          color: MaterialStateProperty.all(clr),
-          onSelectChanged: (value) {
-            if (isDesktop) {
-              showModalSideSheet(
-                  context: context,
-                  ignoreAppBar: true,
-                  width: 640,
-                  barrierColor: Colors.white.withOpacity(barrierOpacity),
-                  // with blur,
-                  barrierDismissible: true,
-                  body: TransactionDetailsDrawer(
-                    transaction: row,
-                  ));
-            } else {
-              context.vRouter.to('/transactions_details/');
-            }
-          },
-          cells: <DataCell>[
-            DataCell(GestureDetector(
-              child: TransactionColumnText(
-                isTransactionTable: true,
-                textTop: isTablet ? row.transactionId.substring(0, 9) : row.transactionId.substring(0, 38),
-                textBottom: "49 ${Strings.secAgo}",
-              ),
-            )),
-            DataCell(TransactionColumnText(
-              isTransactionTable: true,
-              textTop: '${Strings.height}: ${row.block.height}',
-              textBottom: '${Strings.slot}: ${row.block.slot}',
-            )),
-            DataCell(TransactionColumnText(
-              isTransactionTable: true,
-              textTop: row.transactionType.string,
-              textBottom: "",
-              isBottomTextRequired: false,
-            )),
-            const DataCell(TransactionColumnText(
-                isTransactionTable: true, textTop: '3 ${Strings.topl}', textBottom: '44 ${Strings.bobs}')),
-            DataCell(TransactionColumnText(
-              isTransactionTable: true,
-              textTop: '${row.transactionFee} ${Strings.feeAcronym}',
-              textBottom: "",
-              isBottomTextRequired: false,
-            )),
-            DataCell(StatusButton(isTransactionTable: true, status: row.status.string)),
-          ]);
-    } else {
-      return null;
-    }
-  }
-
-  @override
-  bool get isRowCountApproximate => false;
-
-  @override
-  int get rowCount => data.length;
-
-  @override
-  int get selectedRowCount => 0;
-
-  @override
-  void notifyListeners() {
-    super.notifyListeners();
-  }
-}
